requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# ML for usage prediction
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0

# Gemini AI for barcode scanning
google-generativeai>=0.3.0
Pillow>=10.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
